tools/requirements.txt
# Python requirements for ESP32 device configuration tools
paho-mqtt>=1.6.1
esptool>=4.5.1
